Avoid false positive reports from find_regressions
This adds meta_name and bulk_size to the key used to group benchmarks.
Previously it incorrectly compared different benchmarks with each
other, leading to reports with high differences.

(E.g. comparing a benchmark with bulk_size=1000, with a benchmark with
bulk_size=1)

(There seems to be a separate issue in either cr8, or the script
saving the benchmark results into the benchmark table that causes
`bulk_size` to always be NULL)