Stabilize join benchmark by running optimize table
The benchmark

    select * from articles inner join colors on articles.id = colors.id
    where colors.id = -1

depends a lot on the number of segments because one side doesn't match
at all. Our current NL implementation still retries to iterate over the
right side for each left row and if each segment contains no matching
row the cost of setting up the doc-iteration per segment is
proportionally high.

We should optimize this eventually in CrateDB, but this here is to avoid
having the benchmarks fluctuate so much, to reduce false positive
regression notifications.