The compare value doesn't exceed the range of the return type of math functions.
Some fucntions (ceil, round etc) can return an integer. Previously we made
the compare value be a random value of the same type as the column used
as argument for the function call. These functions would yield an integer
which we attempted to compare with a random float literal for eg.
We then attempted to convert the literal to an integer, which could be out
of range for the integer type.
This commit changes the compare value to be of type byte for these functions.