Change query in `in_numeric` to use count(*)
The query became quite slow with 4.0.0 because of the removal of the
implicit HTTP limit. The query returns ~1516033 records which is quite a
lot.

The original purprose of the query was to benchmark the `duration IN`
clause, not the retrieval of large results, so this adjusts the query to
do a `count(*)` which reduces the overhead of the result set processing.